+++ akita/src/api/entity-store.ts
@@ -260,10 +260,30 @@ export class EntityStore<S extends EntityState<E>, E, ActiveEntity = ID> extends
   /**
    * Set the given entity as active.
    */
+<<<<<<< HEAD
   setActive(idOrOptions: S['active'] extends any[] ? S['active'] : (SetActiveOptions | S['active']));
   setActive(idOrOptions: S['active'] | SetActiveOptions) {
     if (Array.isArray(idOrOptions)) {
       this._setActive(idOrOptions);
+=======
+  setActive(idOrOptions: ActiveEntity | SetActiveOptions | null) {
+    let activeId: ActiveEntity;
+
+    if (isObject(idOrOptions)) {
+      if (isNil(this._value().active)) return;
+      (idOrOptions as SetActiveOptions) = Object.assign({ wrap: true }, idOrOptions);
+      const ids = this._value().ids;
+      const currentIdIndex = ids.indexOf(this._value().active);
+      if ((idOrOptions as SetActiveOptions).prev) {
+        const isFirst = currentIdIndex === 0;
+        if (isFirst && !(idOrOptions as SetActiveOptions).wrap) return;
+        activeId = isFirst ? ids[ids.length - 1] : (ids[currentIdIndex - 1] as any);
+      } else if ((idOrOptions as SetActiveOptions).next) {
+        const isLast = ids.length === currentIdIndex + 1;
+        if (isLast && !(idOrOptions as SetActiveOptions).wrap) return;
+        activeId = isLast ? ids[0] : (ids[currentIdIndex + 1] as any);
+      }
+>>>>>>> b6c40f96394d8904158aed9c83e9842760449b95
     } else {
       let activeId: ActiveEntity;
       if (isObject(idOrOptions)) {